Fixed warning

+++ httplib.h
@@ -2862,7 +2862,7 @@ inline bool mmap::open(const char *path) {
   // If the following line doesn't compile due to QuadPart, update Windows SDK.
   // See:
   // https://github.com/yhirose/cpp-httplib/issues/1903#issuecomment-2316520721
-  if (size.QuadPart > std::numeric_limits<decltype(size_)>::max()) {
+  if (static_cast<ULONGLONG>(size.QuadPart) > std::numeric_limits<decltype(size_)>::max()) {
     // `size_t` might be 32-bits, on 32-bits Windows.
     return false;
   }
